changed priority to float to eliminate rounding error

+++ Assets/Scripts/Node.cs
@@ -26,7 +26,7 @@ public class Node : IComparable<Node>
     public float distanceTraveled = Mathf.Infinity;
     public Node previous = null;
 
-    public int priority;
+    public float priority;
 
     public Node(int xIndex, int yIndex, NodeType nodeType)
     {
